fix: test crash root cause (ASan findings + harness hardening)
ASan (RelWithDebInfo -fsanitize=address,undefined) on test_smoke_80
detected stack-use-after-scope in src/game/Commands.cpp:914
inside /setblock handler's block-state parser:

  vector<pair<string_view,string_view>> props;
  string k = ...; string v = ...;
  props.emplace_back(k,v); // dangling: view of locals k/v that die
  // later use of props -> heap corruption -> SIGABRT (timeout 134 Apport)

Fixed by storing props as vector<pair<string,string>> and
building views only from owned strings (sv reserve + emplace).

Also fixed signed-shift UB in ByteBuffer varint/varlong
and Connection::readVarintStream / encrypted len (use uint32/64
intermediate to avoid overflow UB).

Harness hardening:
- TestClient::close() noexcept: close socket before join to wake
  blocking reader, swallow exceptions; destructor noexcept.
- All send helpers (confirmTeleport, sendPosition, chat, dig,
  useItem, moveVehicle, respondKeepAlive, sendPlayerLoadedOnce)
  now guard conn_ null and swallow SocketClosedError -> avoids
  terminate() when server dies mid-send (broken pipe).
- readerLoop guards null conn and checks running_ on timedOut.
- Connection::close() noexcept and destructor noexcept.
- soak_test.py: detect server exit 134/-6 SIGABRT and print
  explicit 'test crashed with SIGABRT (ASan/UBSan?)'.

Verification:
- build-asan: test_spec_wire 328 PASS, test_scoreboard_reset 22,
  test_fuzz 23, test_recipes_mirror 76, test_gameplay_full 244/9,
  test_native ALL PASS, no sanitizer errors (previous run aborted).
- Normal build: spec 328, native ALL PASS, smoke 212 PASS 0 FAIL,
  ctest native/spec/fuzz/recipes 100% PASS, wire 328 maintained.

+++ src/core/ByteBuffer.hpp
@@ -126,23 +126,23 @@ public:
     double f64() { std::uint64_t u = u64(); double f; std::memcpy(&f, &u, 8); return f; }
 
     std::int32_t varint() {
-        std::int32_t result = 0;
+        std::uint32_t result = 0;
         int shift = 0;
         while (true) {
             std::uint8_t b = u8();
-            result |= static_cast<std::int32_t>(b & 0x7F) << shift;
-            if (!(b & 0x80)) return result;
+            result |= static_cast<std::uint32_t>(b & 0x7F) << shift;
+            if (!(b & 0x80)) return static_cast<std::int32_t>(result);
             shift += 7;
             if (shift >= 35) throw std::runtime_error("varint too large");
         }
     }
     std::int64_t varlong() {
-        std::int64_t result = 0;
+        std::uint64_t result = 0;
         int shift = 0;
         while (true) {
             std::uint8_t b = u8();
-            result |= static_cast<std::int64_t>(b & 0x7F) << shift;
-            if (!(b & 0x80)) return result;
+            result |= static_cast<std::uint64_t>(b & 0x7F) << shift;
+            if (!(b & 0x80)) return static_cast<std::int64_t>(result);
             shift += 7;
             if (shift >= 70) throw std::runtime_error("varlong too large");
         }
